feat: merge field data with the current builder form
Refs: #SFB-8
